octan/requirements.txt
colorama==0.4.4
Pillow==9.2.0
term_image==0.4.1
